# api/openapi.yaml
openapi: 3.0.3
info:
  title: DeepWiki-Open API
  description: |
    API documentation for DeepWiki-Open, an AI-powered tool that automatically creates comprehensive, 
    interactive wikis for any GitHub, GitLab, or BitBucket repository.
  version: 1.0.0
  contact:
    name: DeepWiki-Open Team
servers:
  - url: /
    description: Current server
...
paths:
  /api/wiki/{id}:
    get:
      tags:
        - Wiki
      summary: Get wiki content by ID
      description: Retrieves wiki content from cache by ID
      operationId: getWikiById
      parameters:
        - name: id
          in: path
          description: ID of the wiki content to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/WikiStructureModel'
                  - $ref: '#/components/schemas/WikiPage'
        '400':
          description: Wiki ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Wiki content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
...
    WikiPage:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the page
        title:
          type: string
          description: Title of the page
        content:
          type: string
          description: Content of the page
        filePaths:
          type: array
          items:
            type: string
          description: List of file paths related to the page
        importance:
          type: string
          description: Importance level of the page (high, medium, low)
        relatedPages:
          type: array
          items:
            type: string
          description: List of related page IDs
    WikiStructureModel:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the wiki structure
        title:
          type: string
          description: Title of the wiki
        description:
          type: string
          description: Description of the wiki
        pages:
          type: array
          items:
            $ref: '#/components/schemas/WikiPage'
          description: List of wiki pages
...